refactor: Replace LocationPicker with GeoLocationPicker component and enhance API call structure for location search
