+++ public/locale/en.json
@@ -2210,7 +2210,7 @@
   "search_for_allergies_to_add": "Search for allergies to add",
   "search_for_diagnoses_to_add": "Search for diagnoses to add",
   "search_for_facility": "Search for Facility",
-  "search_for_location": "Search for the location",
+  "search_for_location": "Search for a location",
   "search_for_medications_to_add": "Search for medications to add",
   "search_for_symptoms_to_add": "Search for symptoms to add",
   "search_icd11_placeholder": "Search for ICD-11 Diagnoses",
